+++ tasty-plutus/CHANGELOG.md
@@ -4,6 +4,11 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0)
 
 ## Unreleased
 
+### Added
+
+* `paysLovelaceToPubKey`, `paysLovelaceToWallet`.
+* Minting validator support and functions.
+
 ### Changed
 
 * `tagged` is now `datum` to reflect its more general purpose.
@@ -14,6 +19,9 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0)
 
 * `ForCertifying` and `ForRewarding` constructors of `Purpose`. We have no plans
   to support these.
+* `spendsFromSelf`, as self-spending is now handled differently.
+* `compile`, as this is now handled purely internally.
+* `DecodeFailure`, as this condition no longer exists.
 
 ## 2.0 -- 2021-09-14